added support for actions

+++ pyramid_orb/services/model.py
@@ -7,6 +7,9 @@ from pyramid.httpexceptions import HTTPBadRequest, HTTPForbidden, HTTPNotFound
 from pyramid_orb.utils import get_context
 from pyramid_orb.service import OrbService
 
+from ..action import Action
+from ..action import iter_actions
+
 log = logging.getLogger(__name__)
 
 
@@ -31,6 +34,8 @@ class ModelService(OrbService):
         self.record_id = record_id
         self.__record = record
         self.from_collection = from_collection
+
+        self.actions = self.collect_actions_from_model(model)
 
     def __getitem__(self, key):
         schema = self.model.schema()
@@ -91,7 +96,13 @@ class ModelService(OrbService):
         record = self.model(self.record_id, context=context)
         record.update(values)
         record.save()
-        return record.__json__()
+        return record
+
+    def collect_actions_from_model(self, model):
+        actions = {}
+        for action, func in iter_actions(model):
+            actions[action] = func
+        return actions
 
     def get(self):
         values, context = get_context(self.request, model=self.model)
@@ -99,26 +110,43 @@ class ModelService(OrbService):
             return self.model.schema()
         elif self.record_id:
             try:
-                return self.model(self.record_id, context=context)
+                record = self.model(self.record_id, context=context)
             except orb.errors.RecordNotFound:
                 raise HTTPNotFound()
-        else:
-            # convert values to query parameters
-            if values:
-                where = orb.Query.build(values)
-                context.where = where & context.where
-
-            # grab search terms or query
-            search_terms = self.request.params.get('terms') or self.request.params.get('q')
-
-            if search_terms:
-                return self.model.search(search_terms, context=context)
+            action = self.get_record_action()
+            if record is not None and action is not None:
+                return action(record, self.request)
             else:
-                return self.model.select(context=context)
+                return record
+
+        else:
+
+            action = self.get_model_action()
+            if action:
+                return action(self.request)
+
+            else:
+                # convert values to query parameters
+                if values:
+                    where = orb.Query.build(values)
+                    context.where = where & context.where
+
+                # grab search terms or query
+                search_terms = self.request.params.get('terms') or self.request.params.get('q')
+
+                if search_terms:
+                    return self.model.search(search_terms, context=context)
+                else:
+                    return self.model.select(context=context)
 
     def patch(self):
         if self.record_id:
-            return self._update()
+            record = self._update()
+            action = self.get_record_action()
+            if action:
+                return action(record, self.request)
+            else:
+                return record.__json__()
         else:
             raise HTTPBadRequest()
 
@@ -127,12 +155,24 @@ class ModelService(OrbService):
             raise HTTPBadRequest()
         else:
             values, context = get_context(self.request, model=self.model)
-            record = self.model.create(values, context=context)
-            return record.__json__()
+            action = self.get_model_action()
+            if action:
+                # Since we're running a post, there is no record instance
+                # to pass to action. Post actions should always be
+                # classmethods.
+                return action(self.request)
+            else:
+                record = self.model.create(values, context=context)
+                return record.__json__()
 
     def put(self):
         if self.record_id:
-            return self._update()
+            record = self._update()
+            action = self.get_record_action()
+            if action:
+                return action(record, self.request)
+            else:
+                return record
         else:
             raise HTTPBadRequest()
 
@@ -140,11 +180,16 @@ class ModelService(OrbService):
         if self.record_id:
             values, context = get_context(self.request, model=self.model)
             if self.from_collection:
-                return self.from_collection.remove(self.record_id, context=context)
+                return self.from_collection.remove(self.record_id,
+                                                   context=context)
             else:
                 record = self.model(self.record_id, context=context)
-                record.delete()
-                return record
+                action = self.get_record_action()
+                if action:
+                    return action(record, self.request)
+                else:
+                    record.delete()
+                    return record
         else:
             raise HTTPBadRequest()
 
@@ -197,3 +242,20 @@ class ModelService(OrbService):
             output[collector_path + '/{{{0}:id}}'.format(rev_type)] = 'get,delete'
 
         return output
+
+    def get_model_action(self):
+        return self.get_action(True)
+
+    def get_record_action(self):
+        return self.get_action(False)
+
+    def get_action(self, model_action):
+        for key, value in self.request.params.items():
+            if key == 'action':
+                action = Action(name=value,
+                                method=self.request.method.lower(),
+                                model_action=model_action)
+
+                if action in self.actions:
+                    return self.actions[action]
+        return None
